Moved stores outside app/, where it was supposed to be
