+++ stores/colorCheckboxStore.ts
